Rename to cpp-compute-hfd and remove ARGUS references
